requirements.txt
# --- core ---
requests>=2.31.0
PyJWT>=2.8.0
types-requests>=2.31.0
 
# --- test runner ---
pytest>=7.4.0
pytest-html>=4.0.0
 
# --- quality checks ---
ruff>=0.4.0
mypy>=1.8.0